third_party/blink/web_tests/platform/generic/fast/js/mozilla/strict/12.2.1-expected.txt

PASS Function("'use strict'; var eval;") threw exception of type SyntaxError.
PASS !!Function("var eval;") is true
PASS true === true
PASS Function("'use strict'; var x,eval;") threw exception of type SyntaxError.
PASS !!Function("var x,eval;") is true
PASS true === true
PASS Function("'use strict'; var arguments;") threw exception of type SyntaxError.
PASS !!Function("var arguments;") is true
PASS true === true
PASS Function("'use strict'; var x,arguments;") threw exception of type SyntaxError.
PASS !!Function("var x,arguments;") is true
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
